Seperate a include file for unix and windows, add dockerfile, update cleaner script
