+++ docker/Debian_Dockerfile
@@ -5,21 +5,29 @@ FROM debian:jessie
 # need root permission to install some development tools
 USER root
 
-# remove libpcap/libpcap-dev if it already installed
-RUN apt-get remove -y --purge libpcap*
-RUN apt-get autoremove -y
-
-# update local list of software from official reposioty
+# update local list of software from official repository
 RUN apt-get update
 
-# install required tool to build libpcap and jxnet native libarary
-RUN apt-get -y install gcc \
+# install required tools
+RUN apt-get -y install bash \
+        gcc \
         g++ \
-        cmake \
-        bison \
-        flex \
-        git \
-        bash
+        make \
+        git
+
+# required to build libpcap
+RUN apt-get -y install bison \
+        flex
+
+# install autotools
+RUN apt-get -y install automake \
+        autoconf \
+        libtool
+
+# install cmake
+RUN apt-get -y install cmake
+
+RUN apt-get clean
 
 # add Jxnet project into image
 RUN mkdir -p ~/project/
@@ -32,13 +40,20 @@ RUN cd ~/project/Jxnet && \
 # build and install libpcap
 RUN cd ~/project/Jxnet/jxnet-native/libpcap && \
         ./configure && \
-        make && \
-        make install
+        make
 
-# build jxnet native library
+# build jxnet native library with cmake
 RUN cd ~/project/Jxnet/jxnet-native/ && \
         mkdir -p build && \
         cd build && \
         cmake ../ && \
+        make && \
+        make install
+
+# build jxnet native library with autotools
+RUN cd ~/project/Jxnet/jxnet-native/ && \
+        export JAVA_HOME=$(pwd) && \
+        ./bootstrap.sh && \
+        ./configure && \
         make && \
         make install
