+++ jxnet-native/CMakeLists.txt
@@ -81,7 +81,7 @@ if (WIN32 OR MSYS OR MSVC OR MINGW)
 	add_definitions(-DWIN32 -DWIN64 -D__MINGW32__ -D__MINGW64__ -D_WIN32 -D_WIN64)
 	set(CMAKE_LIBRARY_PATH ${CMAKE_CURRENT_SOURCE_DIR}/lib/x64 ${CMAKE_LIBRARY_PATH})
 	LINK_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/lib/x64)
-	include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_SOURCE_DIR}/include/jxnet ${JNI_INCLUDE_DIRS})
+	include_directories(${CMAKE_CURRENT_SOURCE_DIR}/WpdPack/Include ${JNI_INCLUDE_DIRS})
 	message (STATUS "CMAKE_LIBRARY_PATH          = ${CMAKE_LIBRARY_PATH}")
 	find_library(WPCAP_LIB wpcap)
 	find_library(IPHLPAPI_LIB iphlpapi)
@@ -93,7 +93,7 @@ else()
 	set(CMAKE_LIBRARY_PATH ${CMAKE_CURRENT_SOURCE_DIR}/libpcap ${CMAKE_LIBRARY_PATH})
 	LINK_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/libpcap)
 	message (STATUS "CMAKE_LIBRARY_PATH          = ${CMAKE_LIBRARY_PATH}")
-	include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_SOURCE_DIR}/include/jxnet ${JNI_INCLUDE_DIRS})
+	include_directories(${CMAKE_CURRENT_SOURCE_DIR}/libpcap ${JNI_INCLUDE_DIRS})
 	find_library(PCAP_LIB pcap)
 	target_link_libraries(jxnet ${PCAP_LIB})
 	target_link_libraries(application ${PCAP_LIB})
